fix: error caused by restructuring
